Add FS Media retrieval model and controller
The file system read is implemented in Binary in a trait (I wanted to
play with them) and the Media model uses it. Just to keep it consistent
(not because I don't know that Media is already plural) the controller
is named Medias.

Finally I added the routes to make this work.